Moved Spectrum to a common file
